Moved handle_connection into lib.rs, adjusted Use statements to conform with changes
