+++ rust-web-server/src/main.rs
@@ -1,10 +1,4 @@
-use std::{
-    fs,
-    io::{prelude::*, BufReader},
-    net::{TcpListener, TcpStream},
-    thread,
-    time::Duration,
-};
+use std::net::TcpListener;
 
 use rust_web_server::ThreadPool;
 
@@ -16,36 +10,7 @@ fn main() {
         let stream = stream.unwrap();
 
         pool.execute(|| {
-            handle_connection(stream);
+            rust_web_server::handle_connection(stream);
         });
     }
-}
-
-fn handle_connection(mut stream: TcpStream) {
-    let buf_reader = BufReader::new(&stream);
-    let http_request : Vec<_> = buf_reader
-        .lines()
-        .map(|line| line.unwrap())
-        .take_while(|line| !line.is_empty())
-        .collect();
-    println!("Request: {:?}", http_request);
-
-    let request_line = &http_request[0];
-
-    let (status_line, filename) = match request_line.split_whitespace().collect::<Vec<&str>>().as_slice() {
-        ["GET", "/", _] => ("HTTP/1.1 200 OK", "hello.html"),
-        ["GET", "/sleep", _] => {
-            thread::sleep(Duration::from_secs(5));
-            ("HTTP/1.1 200 OK", "hello.html")
-        },
-        _ => ("HTTP/1.1 404 NOT FOUND", "404.html"),
-    };
-    let contents = fs::read_to_string(filename).unwrap();
-    let length = contents.len();
-    let response = format!(
-        "{}\r\nContent-Length: {}\r\n\r\n{}",
-        status_line, length, contents
-    );
-
-    stream.write(response.as_bytes()).unwrap();
 }
